RimWorld: Fix dialog auto-dismiss, MakeThing stuff error, and agent registration
- Fix modal dialog auto-dismiss: move from OnTick to OnUpdate (frame-based)
  since modal dialogs block the tick loop entirely
- Fix FactionNamingDialogPatch: remove static _dismissed flag that prevented
  re-dismissal across saves/sessions
- Fix PlaceBlueprint MakeThing error: ignore stuff parameter for buildings
  that aren't MadeFromStuff (e.g. FueledStove)
- Make agent registration idempotent (return Ok(false) if already registered)
- Fix thread-safe logging: queue log messages for main-thread dispatch
- Fix negative ticksElapsed in episode summary
- Notify all pawns when enabling manual work priorities

+++ dotnet/GameRL.Harmony/Bridge.cs
@@ -590,8 +590,8 @@ namespace GameRL.Harmony
 
                 case EpisodeSummaryMessage m:
                     obj["TotalReward"] = m.TotalReward;
-                    obj["StepCount"] = (long)m.StepCount;
-                    obj["TicksElapsed"] = (long)m.TicksElapsed;
+                    obj["StepCount"] = m.StepCount;
+                    obj["TicksElapsed"] = m.TicksElapsed;
                     obj["RewardBreakdown"] = JToken.FromObject(m.RewardBreakdown ?? new Dictionary<string, double>());
                     if (m.TerminationReason != null)
                         obj["TerminationReason"] = m.TerminationReason;
